reorganize the packages to make one one top-level package as cloudlift (ref: https://github.com/GetSimpl/cloudlift/issues/19)
