+++ cloudlift/config/banner.py
@@ -1,8 +1,3 @@
-import click
-
-from deployment.logging import log_warning
-
-
 def highlight_production():
     print('''\033[91m
 ********************************
