+++ cloudlift/config/region.py
@@ -1,7 +1,7 @@
 import boto3
 
-from config.environment_configuration import EnvironmentConfiguration
-from deployment.logging import log_err
+from cloudlift.config import EnvironmentConfiguration
+from cloudlift.config.logging import log_err
 
 def get_region_for_environment(environment):
     return EnvironmentConfiguration(environment).get_config()[environment]['region']
