+++ cloudlift/deployment/service_template_generator.py
@@ -17,15 +17,15 @@ from troposphere.elasticloadbalancingv2 import (Matcher, RedirectConfig,
                                                 TargetGroupAttribute)
 from troposphere.iam import Role
 
-from config import region as region_service
-from config.account import get_account_id
-from config.decimal_encoder import DecimalEncoder
-from config.stack import get_service_stack_name
-from deployment.deployer import build_config
-from deployment.ecs import DeployAction, EcsClient
-from deployment.logging import log, log_bold
-from deployment.service_information_fetcher import ServiceInformationFetcher
-from deployment.template_generator import TemplateGenerator
+from cloudlift.config import region as region_service
+from cloudlift.config import get_account_id
+from cloudlift.config import DecimalEncoder
+from cloudlift.config import get_service_stack_name
+from cloudlift.deployment.deployer import build_config
+from cloudlift.deployment.ecs import DeployAction, EcsClient
+from cloudlift.config.logging import log, log_bold
+from cloudlift.deployment.service_information_fetcher import ServiceInformationFetcher
+from cloudlift.deployment.template_generator import TemplateGenerator
 
 
 class ServiceTemplateGenerator(TemplateGenerator):
