+++ cloudlift/version/__init__.py
@@ -1,0 +1,1 @@
+VERSION = '1.2.1'
