--- version/__init__.py
@@ -1,1 +1,0 @@
-VERSION = '1.2.0'
